use float precision for math funcs

+++ include/cglm-euler.h
@@ -66,7 +66,7 @@ glm_euler_angles(mat4 m, vec3 dest) {
 
       glm_vec_dup(a[path], dest);
     } else {
-      dest[0] = atan2(m[1][0], m[2][0]);
+      dest[0] = atan2f(m[1][0], m[2][0]);
       dest[1] = M_PI_2;
       dest[2] = 0.0f;
     }
